+++ packages/xrpc/test/src/entities/rate_limit_test.dart
@@ -2,6 +2,9 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+// 🎯 Dart imports:
+import 'dart:io';
+
 // 📦 Package imports:
 import 'package:test/test.dart';
 
@@ -11,8 +14,10 @@ import 'package:xrpc/src/entities/rate_limit.dart';
 void main() {
   group('.fromHeaders', () {
     test('enabled', () {
+      final now = DateTime.now().toUtc().add(Duration(days: 1));
+
       final rateLimit = RateLimit.fromHeaders({
-        'date': 'Wed, 02 Aug 2023 04:27:20 GMT',
+        'date': HttpDate.format(now),
         'RateLimit-Limit': '1000',
         'RateLimit-Remaining': '0',
         'RateLimit-Reset': '50',
@@ -21,7 +26,6 @@ void main() {
 
       expect(rateLimit.limitCount, 1000);
       expect(rateLimit.remainingCount, 0);
-      expect(rateLimit.resetAt.toIso8601String(), '2023-08-02T04:28:10.000Z');
       expect(rateLimit.policy.limitCount, 100);
       expect(rateLimit.policy.window.inSeconds, 300);
 
@@ -29,9 +33,11 @@ void main() {
       expect(rateLimit.isNotExceeded, isFalse);
     });
 
-    test('enabled and not exceeded', () {
+    test('enabled and not exceeded due to remaining', () {
+      final now = DateTime.now().toUtc().add(Duration(days: 1));
+
       final rateLimit = RateLimit.fromHeaders({
-        'date': 'Wed, 02 Aug 2023 04:27:20 GMT',
+        'date': HttpDate.format(now),
         'RateLimit-Limit': '1000',
         'RateLimit-Remaining': '1',
         'RateLimit-Reset': '50',
@@ -40,7 +46,26 @@ void main() {
 
       expect(rateLimit.limitCount, 1000);
       expect(rateLimit.remainingCount, 1);
-      expect(rateLimit.resetAt.toIso8601String(), '2023-08-02T04:28:10.000Z');
+      expect(rateLimit.policy.limitCount, 100);
+      expect(rateLimit.policy.window.inSeconds, 300);
+
+      expect(rateLimit.isExceeded, isFalse);
+      expect(rateLimit.isNotExceeded, isTrue);
+    });
+
+    test('enabled and not exceeded due to resetAt (past)', () {
+      final dayAgo = DateTime.now().toUtc().add(Duration(days: -1));
+
+      final rateLimit = RateLimit.fromHeaders({
+        'date': HttpDate.format(dayAgo),
+        'RateLimit-Limit': '1000',
+        'RateLimit-Remaining': '0',
+        'RateLimit-Reset': '50',
+        'RateLimit-Policy': '100;w=300',
+      });
+
+      expect(rateLimit.limitCount, 1000);
+      expect(rateLimit.remainingCount, 0);
       expect(rateLimit.policy.limitCount, 100);
       expect(rateLimit.policy.window.inSeconds, 300);
 
@@ -61,6 +86,60 @@ void main() {
 
       expect(rateLimit.isExceeded, isFalse);
       expect(rateLimit.isNotExceeded, isTrue);
+    });
+  });
+
+  group('.waitUntilWait', () {
+    test('when need to wait', () async {
+      final now = DateTime.now().toUtc();
+
+      final rateLimit = RateLimit.fromHeaders({
+        'date': HttpDate.format(now),
+        'RateLimit-Limit': '1000',
+        'RateLimit-Remaining': '0',
+        //! Assumes a margin of error of a few milliseconds.
+        'RateLimit-Reset': '11',
+        'RateLimit-Policy': '100;w=300',
+      });
+
+      final result = await rateLimit.waitUntilReset();
+
+      final waitedInSeconds = DateTime.now().toUtc().difference(now).inSeconds;
+      expect(waitedInSeconds, 10); //! About 10998 milliseconds.
+
+      expect(result, isTrue);
+    });
+
+    test('when not need to wait due to remaining', () async {
+      final now = DateTime.now().toUtc();
+
+      final rateLimit = RateLimit.fromHeaders({
+        'date': HttpDate.format(now),
+        'RateLimit-Limit': '1000',
+        'RateLimit-Remaining': '1',
+        'RateLimit-Reset': '10',
+        'RateLimit-Policy': '100;w=300',
+      });
+
+      final result = await rateLimit.waitUntilReset();
+
+      expect(result, isFalse);
+    });
+
+    test('when not need to wait due to resetAt (past)', () async {
+      final now = DateTime.now().toUtc().add(Duration(days: -1));
+
+      final rateLimit = RateLimit.fromHeaders({
+        'date': HttpDate.format(now),
+        'RateLimit-Limit': '1000',
+        'RateLimit-Remaining': '0',
+        'RateLimit-Reset': '10',
+        'RateLimit-Policy': '100;w=300',
+      });
+
+      final result = await rateLimit.waitUntilReset();
+
+      expect(result, isFalse);
     });
   });
 }