+++ packages/xrpc/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.2
 
 - Supported `RateLimit`. You can access this property from `XRPCResponse.rateLimit`. If there is no rate limits, then it's always null. ([#636](https://github.com/myConsciousness/atproto.dart/issues/636))
+- Added `.waitUntilReset` method on `RateLimit` ([#643](https://github.com/myConsciousness/atproto.dart/issues/643))
 
 ## v0.2.1